Add files via upload
A non-invasive method of measuring bioimpedance is presented to study the effect of moisturizing cream on human skin. A novel two-electrode system is developed to get electrical access at the area of interest on human skin. Controlled experiments are conducted for a duration of 47 days. Bioimpedance is measured by LCR meter in the frequency range of 1 Hz to 5 kHz. The measured impedance is then fitted with the Cole model and the relevant Cole parameters are extracted. Analysis of variance is performed on the Cole parameters to gain statistical relevance. It is found that one of the Cole parameters, namely the relaxation-time changes significantly due to the treatment of the skin with a moisturizing cream at more than 95% confidence level. The relaxation-time increases sharply during the treatment period and drops sharply in the post-treatment period. The effectiveness of the two types of cream is studied. The gender distinguishability during the application of a moisturizing cream is also studied. A very generic bioimpedance measurement method to assess the physiological condition of local skin is demonstrated here and the method has potential application in human skin related diseases and treatments.